+++ google/cloud/pubsub/internal/streaming_subscription_batch_source_test.cc
@@ -130,8 +130,8 @@ class FakeStream {
   std::deque<promise<bool>> actions_;
 };
 
-pubsub::SubscriptionOptions TestSubscriptionOptions() {
-  return pubsub::SubscriptionOptions()
+pubsub::SubscriberOptions TestSubscriptionOptions() {
+  return pubsub::SubscriberOptions()
       .set_max_outstanding_messages(100)
       .set_max_outstanding_bytes(100 * 1024 * 1024L)
       .set_max_deadline_time(std::chrono::seconds(300));